+++ src/anschluss/pcf8574.rs
@@ -178,7 +178,7 @@ pub struct InVerwendung {
     pub port: kleiner_8,
 }
 
-#[derive(Debug, Clone, Copy, zugkontrolle_derive::From)]
+#[derive(Debug, Clone, Copy, zugkontrolle_macros::From)]
 #[allow(variant_size_differences)]
 pub enum ReservierenFehler {
     Deaktiviert(Deaktiviert),