+++ src/application/gleis/gerade.rs
@@ -3,7 +3,7 @@
 use std::{fmt::Debug, hash::Hash};
 
 use serde::{Deserialize, Serialize};
-use zugkontrolle_derive::alias_serialisiert_unit;
+use zugkontrolle_macros::alias_serialisiert_unit;
 
 use crate::{
     application::{typen::*, verbindung::Verbindung},