+++ src/application/gleis/kreuzung.rs
@@ -3,7 +3,7 @@
 use std::fmt::Debug;
 
 use serde::{Deserialize, Serialize};
-use zugkontrolle_derive::alias_serialisiert_unit;
+use zugkontrolle_macros::alias_serialisiert_unit;
 
 pub use crate::application::gleis::weiche::gerade::{
     Richtung, RichtungAnschlüsse, RichtungAnschlüsseSerialisiert,