+++ src/application/typen/rechteck.rs
@@ -27,7 +27,7 @@ impl Rechteck {
     }
 
     /// Verschiebe das Rechteck um Vektor.
-    #[zugkontrolle_derive::chain]
+    #[zugkontrolle_macros::chain]
     pub fn verschiebe(&mut self, bewegung: &Vektor) {
         self.ecke_a += bewegung;
         self.ecke_b += bewegung;
@@ -41,7 +41,7 @@ impl Rechteck {
 
     /// Dehne das Rechteck aus, so dass es um `winkel`-Rotation um `(0, 0)` (im Uhrzeigersinn)
     /// in das angepasste (nicht rotierte) Rechteck passt.
-    #[zugkontrolle_derive::chain]
+    #[zugkontrolle_macros::chain]
     pub fn respektiere_rotation(&mut self, winkel: &Winkel) {
         let Rechteck { mut ecke_a, mut ecke_b } = *self;
         let mut ecke_c = Vektor { x: ecke_a.x, y: ecke_b.y };